1110
파일없는 게시글 삭제 코드 수정

+++ src/main/java/org/zerock/service/board/BoardService.java
@@ -150,8 +150,10 @@ public class BoardService {
 		
 		File[] listFiles = folder.listFiles();
 		
-		for (File file : listFiles) {
-			file.delete();
+		if (listFiles != null) {
+			for (File file : listFiles) {
+				file.delete();
+			}
 		}
 		
 		folder.delete();
